requirements.txt
streamlit==1.28.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
reportlab==4.0.7
python-dotenv==1.0.0
pandas==2.1.3
Pillow==10.1.0
openpyxl==3.1.2
xlrd==2.0.1
